Add skeleton of deep work cheatsheet
This might not actually be worth it.  I thought that there was a nice,
hierarchical structure to most of the book, but the first chapter
doesn't look like it'll translate well to an outline.  Perhaps the later
chapters will work better--and the rules chapters are what I'm really
after--but I'm giving up for now.